[MOV] core, *: move intervals to tools
The intervals are the generalized concept that can be applied to any model with
start and stop date.

This commit moves the `Interval` class and a few utils methods from resource and
appointment to odoo/tools. It improve flexibility and reusability across different models without
any additional dependencies.

Task-3997545
