+++ addons/hr_contract/models/hr_employee.py
@@ -8,8 +8,8 @@ from dateutil.relativedelta import relativedelta
 
 from odoo import _, api, fields, models
 from odoo.osv import expression
-from odoo.addons.resource.models.utils import Intervals
 from odoo.exceptions import UserError
+from odoo.tools.date_intervals import Intervals
 
 
 class HrEmployeePublic(models.Model):
