+++ addons/hr_contract/tests/test_resource.py
@@ -3,8 +3,8 @@
 from datetime import date, datetime
 from pytz import utc, timezone
 
-from odoo.addons.resource.models.utils import Intervals, sum_intervals
 from odoo.fields import Date
+from odoo.tools.date_intervals import Intervals, sum_intervals
 
 from .common import TestContractCommon
