+++ addons/hr_work_entry_contract/models/hr_contract.py
@@ -9,9 +9,9 @@ import pytz
 from dateutil.relativedelta import relativedelta
 
 from odoo import api, fields, models, _
-from odoo.addons.resource.models.utils import string_to_datetime, Intervals
 from odoo.osv import expression
 from odoo.tools import ormcache, format_list
+from odoo.tools.date_intervals import string_to_datetime
 from odoo.exceptions import UserError
 
 from .hr_work_intervals import WorkIntervals
