+++ addons/l10n_fr_hr_work_entry_holidays/models/hr_contract.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 import pytz
 
 from odoo import models
-from odoo.addons.resource.models.utils import datetime_to_string
+from odoo.tools.date_intervals import datetime_to_string
 
 
 class HrContract(models.Model):
