+++ addons/resource/models/resource_mixin.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 from pytz import utc
 
 from odoo import api, fields, models
-from .utils import timezone_datetime
+from odoo.tools.date_intervals import timezone_datetime
 
 
 class ResourceMixin(models.AbstractModel):
