+++ addons/website_event_exhibitor/models/event_sponsor.py
@@ -5,8 +5,8 @@ from datetime import datetime, timedelta
 from pytz import timezone, utc
 
 from odoo import api, fields, models, _
-from odoo.addons.resource.models.utils import float_to_time
 from odoo.tools import is_html_empty
+from odoo.tools.date_intervals import float_to_time
 from odoo.tools.translate import html_translate
