+++ odoo/addons/base/tests/__init__.py
@@ -7,6 +7,7 @@ from . import test_barcode
 from . import test_base
 from . import test_basecase
 from . import test_cache
+from . import test_date_intervals
 from . import test_date_utils
 from . import test_deprecation
 from . import test_db_cursor
